+++ scratch.js
@@ -6,32 +6,24 @@ spacetime.extend(require('./plugins/week-math/plugin.js'))
 //
 //
 
-// let d = spacetime('2018-10-01')
-// d = d.startOf('quarter')
-// d = d.add(1, 'quarter')
-// d = d.endOf('quarter')
+// let year = 2020
+// let start = spacetime('dec 1 ' + year)
+// let end = spacetime('jan 20 ' + (year + 1))
+// start.every(end, 'week').forEach((d) => {
+//   console.log(d.format(), d.week())
+// })
 
-// let d = spacetime('2020-10-01')
-// d = d.endOf('quarter')
-// console.log(d.format())
+console.log(spacetime().isSame('year', spacetime.now()))
 
-// let d = spacetime('2018-11-01')
-// // // d = d.startOf('quarter').add(1, 'quarter')
-// d = d.endOf('quarter')
-// console.log(d.format())
+// bug:
+// console.log(spacetime('2022-01-03').week())
 
-let d = spacetime('2019-10-01')
-d = d.minus(8, 'quarter')
-console.log(d.format(), 'end')
-
-// let d = spacetime('2017-03-01')
-// d = d.minus(1, 'quarter')
-// d = d.minus(1, 'quarter')
-// d = d.minus(1, 'quarter')
-// d = d.minus(1, 'quarter')
-// console.log(d.format())
-// d = d.add(1, 'quarter')
-// d = d.add(1, 'quarter')
-// d = d.add(1, 'quarter')
-// d = d.add(1, 'quarter')
-// console.log(d.format())
+/*
+2020-12-07 50
+2020-12-14 51
+2020-12-21 52
+2020-12-28 53
+2021-01-04 2
+2021-01-11 3
+2021-01-18 4
+*/